fix: remove redundant consts, add custom hook
